Move team functions from Dex to Teams
