+++ sim/battle-stream.ts
@@ -10,6 +10,7 @@
  */
 
 import {Streams, Utils} from '../lib';
+import {Teams} from './teams';
 import {Battle} from './battle';
 
 /**
@@ -215,7 +216,7 @@ export class BattleStream extends Streams.ObjectReadWriteStream<string> {
 				throw new Error(`Team requested for slot ${message}, but that slot does not exist.`);
 			}
 			const side = this.battle!.sides[slotNum];
-			const team = Dex.packTeam(side.team);
+			const team = Teams.pack(side.team);
 			this.push(`requesteddata\n${team}`);
 			break;
 		case 'version':
