+++ sim/dex-data.ts
@@ -302,115 +302,3 @@ export function combine(obj: AnyObject, ...data: (AnyObject | null)[]): AnyObjec
 	}
 	return obj;
 }
-
-// export class PokemonSet {
-// 	/**
-// 	 * The Pokemon's set's nickname, which is identical to its base
-// 	 * species if not specified by the player, e.g. "Minior".
-// 	 */
-// 	name: string;
-// 	/**
-// 	 * The Pokemon's species, e.g. "Minior-Red".
-// 	 * This should always be converted to an id before use.
-// 	 */
-// 	species: string;
-// 	/**
-// 	 * The Pokemon's set's item. This can be an id, e.g. "whiteherb"
-// 	 * or a full name, e.g. "White Herb".
-// 	 * This should always be converted to an id before use.
-// 	 */
-// 	item: string;
-// 	/**
-// 	 * The Pokemon's set's ability. This can be an id, e.g. "shieldsdown"
-// 	 * or a full name, e.g. "Shields Down".
-// 	 * This should always be converted to an id before use.
-// 	 */
-// 	ability: string;
-// 	/**
-// 	 * An array of the Pokemon's set's moves. Each move can be an id,
-// 	 * e.g. "shellsmash" or a full name, e.g. "Shell Smash"
-// 	 * These should always be converted to ids before use.
-// 	 */
-// 	moves: string[];
-// 	/**
-// 	 * The Pokemon's set's nature. This can be an id, e.g. "adamant"
-// 	 * or a full name, e.g. "Adamant".
-// 	 * This should always be converted to an id before use.
-// 	 */
-// 	nature: string;
-// 	/**
-// 	 * The Pokemon's set's gender.
-// 	 */
-// 	gender: GenderName;
-// 	/**
-// 	 * The Pokemon's set's effort values, used in stat calculation.
-// 	 * These must be between 0 and 255, inclusive.
-// 	 */
-// 	evs: StatsTable;
-// 	/**
-// 	 * The Pokemon's individual values, used in stat calculation.
-// 	 * These must be between 0 and 31, inclusive.
-// 	 * These are also used as DVs, or determinant values, in Gens
-// 	 * 1 and 2, which are represented as even numbers from 0 to 30.
-// 	 * From Gen 2 and on, IVs/DVs are used to determine Hidden Power's
-// 	 * type, although in Gen 7 a Pokemon may be legally altered such
-// 	 * that its stats are calculated as if these values were 31 via
-// 	 * Bottlecaps. Currently, PS handles this by considering any
-// 	 * IV of 31 in Gen 7 to count as either even or odd for the purpose
-// 	 * of validating a Hidden Power type, though restrictions on
-// 	 * possible IVs for event-only Pokemon are still considered.
-// 	 */
-// 	ivs: StatsTable;
-// 	/**
-// 	 * The Pokemon's level. This is usually between 1 and 100, inclusive,
-// 	 * but the simulator supports levels up to 9999 for testing purposes.
-// 	 */
-// 	level: number;
-// 	/**
-// 	 * Whether the Pokemon is shiny or not. While having no direct
-// 	 * competitive effect except in a few OMs, certain Pokemon cannot
-// 	 * be legally obtained as shiny, either as a whole or with certain
-// 	 * event-only abilities or moves.
-// 	 */
-// 	shiny?: boolean;
-// 	/**
-// 	 * The Pokemon's set's happiness value. This is used only for
-// 	 * calculating the base power of the moves Return and Frustration.
-// 	 * This value must be between 0 and 255, inclusive.
-// 	 */
-// 	happiness: number;
-// 	/**
-// 	 * The Pokemon's set's Hidden Power type. This value is intended
-// 	 * to be used to manually set a set's HP type in Gen 7 where
-// 	 * its IVs do not necessarily reflect the user's intended type.
-// 	 * TODO: actually support this in the teambuilder.
-// 	 */
-// 	hpType?: string;
-// 	/**
-// 	 * The pokeball this Pokemon is in. Like shininess, this property
-// 	 * has no direct competitive effects, but has implications for
-// 	 * event legality. For example, any Rayquaza that knows V-Create
-// 	 * must be sent out from a Cherish Ball.
-// 	 * TODO: actually support this in the validator, switching animations,
-// 	 * and the teambuilder.
-// 	 */
-// 	pokeball?: string;
-//
-// 	constructor(data: Partial<PokemonSet>) {
-// 		this.name = '';
-// 		this.species = '';
-// 		this.item = '';
-// 		this.ability = 'noability';
-// 		this.moves = [];
-// 		this.nature = '';
-// 		this.gender = '';
-// 		this.evs = {hp: 0, atk: 0, def: 0, spa: 0, spd: 0, spe: 0};
-// 		this.ivs = {hp: 31, atk: 31, def: 31, spa: 31, spd: 31, spe: 31};
-// 		this.level = 100;
-// 		this.shiny = undefined;
-// 		this.happiness = 255; // :)
-// 		this.hpType = undefined;
-// 		this.pokeball = undefined;
-// 		Object.assign(this, data);
-// 	}
-// }
