+++ sim/dex.ts
@@ -36,7 +36,6 @@ import {Item, DexItems} from './dex-items';
 import {Ability, DexAbilities} from './dex-abilities';
 import {Species, DexSpecies} from './dex-species';
 import {Format, DexFormats} from './dex-formats';
-import {PRNG, PRNGSeed} from './prng';
 import {Utils} from '../lib';
 
 const BASE_MOD = 'gen8' as ID;
@@ -357,15 +356,6 @@ export class ModdedDex {
 		return num >>> 0;
 	}
 
-	getTeamGenerator(format: Format | string, seed: PRNG | PRNGSeed | null = null) {
-		const TeamGenerator = require(dexes['base'].forFormat(format).dataDir + '/random-teams').default;
-		return new TeamGenerator(format, seed);
-	}
-
-	generateTeam(format: Format | string, options: PlayerOptions | null = null): PokemonSet[] {
-		return this.getTeamGenerator(format, options?.seed).getTeam(options);
-	}
-
 	dataSearch(
 		target: string, searchIn?: ('Pokedex' | 'Moves' | 'Abilities' | 'Items' | 'Natures')[] | null, isInexact?: boolean
 	): AnyObject[] | null {
@@ -421,256 +411,6 @@ export class ModdedDex {
 		}
 
 		return searchResults;
-	}
-
-	packTeam(team: PokemonSet[] | null): string {
-		if (!team) return '';
-
-		function getIv(ivs: StatsTable, s: keyof StatsTable): string {
-			return ivs[s] === 31 || ivs[s] === undefined ? '' : ivs[s].toString();
-		}
-
-		let buf = '';
-		for (const set of team) {
-			if (buf) buf += ']';
-
-			// name
-			buf += (set.name || set.species);
-
-			// species
-			const id = toID(set.species || set.name);
-			buf += '|' + (toID(set.name || set.species) === id ? '' : id);
-
-			// item
-			buf += '|' + toID(set.item);
-
-			// ability
-			buf += '|' + toID(set.ability);
-
-			// moves
-			buf += '|' + set.moves.map(toID).join(',');
-
-			// nature
-			buf += '|' + (set.nature || '');
-
-			// evs
-			let evs = '|';
-			if (set.evs) {
-				evs = '|' + (set.evs['hp'] || '') + ',' + (set.evs['atk'] || '') + ',' + (set.evs['def'] || '') + ',' + (set.evs['spa'] || '') + ',' + (set.evs['spd'] || '') + ',' + (set.evs['spe'] || '');
-			}
-			if (evs === '|,,,,,') {
-				buf += '|';
-			} else {
-				buf += evs;
-			}
-
-			// gender
-			if (set.gender) {
-				buf += '|' + set.gender;
-			} else {
-				buf += '|';
-			}
-
-			// ivs
-			let ivs = '|';
-			if (set.ivs) {
-				ivs = '|' + getIv(set.ivs, 'hp') + ',' + getIv(set.ivs, 'atk') + ',' + getIv(set.ivs, 'def') +
-					',' + getIv(set.ivs, 'spa') + ',' + getIv(set.ivs, 'spd') + ',' + getIv(set.ivs, 'spe');
-			}
-			if (ivs === '|,,,,,') {
-				buf += '|';
-			} else {
-				buf += ivs;
-			}
-
-			// shiny
-			if (set.shiny) {
-				buf += '|S';
-			} else {
-				buf += '|';
-			}
-
-			// level
-			if (set.level && set.level !== 100) {
-				buf += '|' + set.level;
-			} else {
-				buf += '|';
-			}
-
-			// happiness
-			if (set.happiness !== undefined && set.happiness !== 255) {
-				buf += '|' + set.happiness;
-			} else {
-				buf += '|';
-			}
-
-			if (set.pokeball || set.hpType || set.gigantamax) {
-				buf += ',' + (set.hpType || '');
-				buf += ',' + toID(set.pokeball || '');
-				buf += ',' + (set.gigantamax ? 'G' : '');
-			}
-		}
-
-		return buf;
-	}
-
-	fastUnpackTeam(buf: string): PokemonSet[] | null {
-		if (!buf) return null;
-		if (typeof buf !== 'string') return buf;
-		if (buf.startsWith('[') && buf.endsWith(']')) {
-			buf = this.packTeam(JSON.parse(buf));
-		}
-
-		const team = [];
-		let i = 0;
-		let j = 0;
-
-		// limit to 24
-		for (let count = 0; count < 24; count++) {
-			const set: PokemonSet = {} as PokemonSet;
-			team.push(set);
-
-			// name
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			set.name = buf.substring(i, j);
-			i = j + 1;
-
-			// species
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			set.species = buf.substring(i, j) || set.name;
-			i = j + 1;
-
-			// item
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			set.item = buf.substring(i, j);
-			i = j + 1;
-
-			// ability
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			const ability = buf.substring(i, j);
-			const species = dexes['base'].species.get(set.species);
-			set.ability = ['', '0', '1', 'H', 'S'].includes(ability) ?
-				species.abilities[ability as '0' || '0'] || (ability === '' ? '' : '!!!ERROR!!!') :
-				ability;
-			i = j + 1;
-
-			// moves
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			set.moves = buf.substring(i, j).split(',', 24).filter(x => x);
-			i = j + 1;
-
-			// nature
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			set.nature = buf.substring(i, j);
-			i = j + 1;
-
-			// evs
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			if (j !== i) {
-				const evs = buf.substring(i, j).split(',', 6);
-				set.evs = {
-					hp: Number(evs[0]) || 0,
-					atk: Number(evs[1]) || 0,
-					def: Number(evs[2]) || 0,
-					spa: Number(evs[3]) || 0,
-					spd: Number(evs[4]) || 0,
-					spe: Number(evs[5]) || 0,
-				};
-			}
-			i = j + 1;
-
-			// gender
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			if (i !== j) set.gender = buf.substring(i, j);
-			i = j + 1;
-
-			// ivs
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			if (j !== i) {
-				const ivs = buf.substring(i, j).split(',', 6);
-				set.ivs = {
-					hp: ivs[0] === '' ? 31 : Number(ivs[0]) || 0,
-					atk: ivs[1] === '' ? 31 : Number(ivs[1]) || 0,
-					def: ivs[2] === '' ? 31 : Number(ivs[2]) || 0,
-					spa: ivs[3] === '' ? 31 : Number(ivs[3]) || 0,
-					spd: ivs[4] === '' ? 31 : Number(ivs[4]) || 0,
-					spe: ivs[5] === '' ? 31 : Number(ivs[5]) || 0,
-				};
-			}
-			i = j + 1;
-
-			// shiny
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			if (i !== j) set.shiny = true;
-			i = j + 1;
-
-			// level
-			j = buf.indexOf('|', i);
-			if (j < 0) return null;
-			if (i !== j) set.level = parseInt(buf.substring(i, j));
-			i = j + 1;
-
-			// happiness
-			j = buf.indexOf(']', i);
-			let misc;
-			if (j < 0) {
-				if (i < buf.length) misc = buf.substring(i).split(',', 4);
-			} else {
-				if (i !== j) misc = buf.substring(i, j).split(',', 4);
-			}
-			if (misc) {
-				set.happiness = (misc[0] ? Number(misc[0]) : 255);
-				set.hpType = misc[1] || '';
-				set.pokeball = this.toID(misc[2] || '');
-				set.gigantamax = !!misc[3];
-			}
-			if (j < 0) break;
-			i = j + 1;
-		}
-
-		return team;
-	}
-
-	/**
-	* Use instead of Dex.packTeam to generate more human-readable team output.
-	*/
-	stringifyTeam(team: PokemonSet[], nicknames?: string[], hideStats?: boolean) {
-		let output = '';
-		for (const [i, mon] of team.entries()) {
-			const species = Dex.species.get(mon.species);
-			const nickname = nicknames?.[i];
-			output += nickname && nickname !== species.baseSpecies ? `${nickname} (${species.name})` : species.name;
-			output += mon.item ? ` @ ${Dex.items.get(mon.item).name}<br />` : `<br />`;
-			output += `Ability: ${Dex.abilities.get(mon.ability).name}<br />`;
-			if (typeof mon.happiness === 'number' && mon.happiness !== 255) output += `Happiness: ${mon.happiness}<br />`;
-			if (mon.gigantamax) output += `Gigantamax: Yes<br />`;
-			if (!hideStats) {
-				const evs = [];
-				for (const stat in mon.evs) {
-					if (mon.evs[stat as StatID]) evs.push(`${mon.evs[stat as StatID]} ${stat}`);
-				}
-				if (evs.length) output += `EVs: ${evs.join(' / ')}<br />`;
-				if (mon.nature) output += `${this.natures.get(mon.nature).name} Nature<br />`;
-				const ivs = [];
-				for (const stat in mon.ivs) {
-					if (mon.ivs[stat as StatID] !== 31) ivs.push(`${mon.ivs[stat as StatID]} ${stat}`);
-				}
-				if (ivs.length) output += `IVs: ${ivs.join(' / ')}<br />`;
-			}
-			output += mon.moves.map(move => `- ${Dex.moves.get(move).name}<br />`).join('');
-			output += '<br />';
-		}
-		return output;
 	}
 
 	loadDataFile(basePath: string, dataType: DataType | 'Aliases'): AnyObject {
@@ -833,3 +573,5 @@ export namespace Dex {
 	export type SecondaryEffect = import('./dex-moves').SecondaryEffect;
 	export type RuleTable = import('./dex-formats').RuleTable;
 }
+
+export default Dex;
