+++ sim/examples/battle-stream-example.ts
@@ -9,8 +9,7 @@
  * @author Guangcong Luo <guangcongluo@gmail.com>
  */
 
-import {BattleStream, getPlayerStreams} from '../battle-stream';
-import {Dex} from '../dex';
+import {BattleStream, getPlayerStreams, Teams} from '..';
 import {RandomPlayerAI} from '../tools/random-player-ai';
 
 /*********************************************************************
@@ -24,11 +23,11 @@ const spec = {
 };
 const p1spec = {
 	name: "Bot 1",
-	team: Dex.packTeam(Dex.generateTeam('gen7randombattle')),
+	team: Teams.pack(Teams.generate('gen7randombattle')),
 };
 const p2spec = {
 	name: "Bot 2",
-	team: Dex.packTeam(Dex.generateTeam('gen7randombattle')),
+	team: Teams.pack(Teams.generate('gen7randombattle')),
 };
 
 const p1 = new RandomPlayerAI(streams.p1);
